Move executor code id table to the app
